backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
aiosqlite>=0.20.0
httpx>=0.28.0
feedparser>=6.0.11
anthropic>=0.79.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
apscheduler>=3.10.4
python-dotenv>=1.0.1
